requirements.txt
# Core Django stack
Django==5.0.8
djangorestframework==3.15.2
django-filter==24.2
django-cors-headers==4.4.0
django-appconf==1.1.0
django-imagekit==5.0.0
sqlparse==0.5.3
pytz==2025.2
tzdata==2025.2

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.10.1

# Configuration
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Celery stack
celery==5.3.1
kombu==5.5.4
billiard==4.2.1
vine==5.1.0
amqp==5.3.1

# Image processing
Pillow==10.1.0
pilkit==3.0

# API Documentation
drf-spectacular==0.27.2

# Production
gunicorn==22.0.0
whitenoise==6.7.0

# Development & Testing
coverage==7.10.6
pytest-django==4.8.0
factory-boy==3.3.0
prompt_toolkit==3.0.52
colorama==0.4.6
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
wcwidth==0.2.13

# Networking & Utilities
requests==2.31.0
urllib3==2.5.0
charset-normalizer==3.4.2
idna==3.10
certifi==2025.6.15
packaging==25.0
python-dateutil==2.9.0.post0
six==1.17.0
